etl-containers/requirements.txt
kafka-python
pandas
pyarrow
pyspark
sqlalchemy
psycopg2-binary
boto3
